final update of TravelManagementService

+++ DTO/Drone.cs
@@ -2,5 +2,5 @@ namespace ChallengeCode.DTO;
 
 public class Drone{
     public string Name { get; set; }
-    public double MaximumWeight { get; set; }
+    public int MaximumWeight { get; set; }
 }
