+++ DTO/Location.cs
@@ -2,5 +2,5 @@ namespace ChallengeCode.DTO;
 
 public class Location{
     public string Name { get; set; }
-    public double PackageWeight { get; set; }
+    public int PackageWeight { get; set; }
 }
